PMM-1079-11544 Refactor.
