+++ agent/runner/actions/common_test.go
@@ -1,0 +1,28 @@
+// Copyright 2019 Percona LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package actions provides Actions implementations.
+package actions
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPrepareQueryWithDatabaseTableName(t *testing.T) {
+	assert.Equal(t, "SHOW /* pmm-agent */ INDEX IN `db`.`table`", prepareQueryWithDatabaseTableName("SHOW /* pmm-agent */ INDEX IN", "db.table"))
+	assert.Equal(t, "SHOW /* pmm-agent */ INDEX IN \"`db`.`table`\"", prepareQueryWithDatabaseTableName("SHOW /* pmm-agent */ INDEX IN", "`db.table`"))
+	assert.Equal(t, "SHOW /* pmm-agent */ INDEX IN \"`db`.`table`\"", prepareQueryWithDatabaseTableName("SHOW /* pmm-agent */ INDEX IN", "`db`.`table`"))
+}
