+++ agent/runner/actions/mysql_show_index_action.go
@@ -16,8 +16,6 @@ package actions
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"time"
 
 	"github.com/percona/pmm/agent/tlshelpers"
@@ -65,12 +63,7 @@ func (a *mysqlShowIndexAction) Run(ctx context.Context) ([]byte, error) {
 	defer db.Close() //nolint:errcheck
 	defer tlshelpers.DeregisterMySQLCerts()
 
-	// use %#q to convert "table" to `"table"` and `table` to "`table`" to avoid SQL injections
-	query := fmt.Sprintf("SHOW /* pmm-agent */ INDEX IN %#q", a.params.Table)
-	// handle case when there is table name together with database name
-	query = strings.ReplaceAll(query, ".", "`.`")
-
-	rows, err := db.QueryContext(ctx, query)
+	rows, err := db.QueryContext(ctx, prepareQueryWithDatabaseTableName("SHOW /* pmm-agent */ INDEX IN", a.params.Table))
 	if err != nil {
 		return nil, err
 	}
